backend/api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-cors==0.8.1
aiokafka==0.12.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
bcrypt==3.2.0
beartype==0.15.0
bleach==6.3.0
bokeh==3.8.1
boto3==1.40.70
botocore==1.40.70
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
contourpy==1.3.3
decorator==5.2.1
deltalake==0.17.4
diskcache==5.6.3
dnspython==2.8.0
docker==7.1.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.121.1
frozenlist==1.8.0
fs==2.4.16
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-bigquery==3.29.0
google-cloud-core==2.5.0
google-cloud-pubsub==2.33.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h3==4.3.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jupyter_bokeh==4.0.5
jupyterlab_widgets==3.0.16
linkify-it-py==2.0.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
motor==3.7.1
multidict==6.7.0
narwhals==2.11.0
networkx==3.5
numpy==2.3.4
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==2.3.3
panel==1.8.3
param==2.2.1
parso==0.8.5
passlib==1.7.4
pathway==0.26.4
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.3
sqlalchemy==2.0.44
asyncpg==0.31.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
python-sat==1.8.dev24
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.49.3
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
xyzservices==2025.10.0
yarl==1.22.0
zipp==3.23.0
lz4
cramjam
twilio
greenlet
slack_sdk>=3.21.0
anthropic==0.75.0